+++ foxes-and-rabbits-handout/Rhino.java
@@ -24,7 +24,7 @@ public class Rhino extends Animal
     // A shared random number generator to control breeding.
     private static final Random rand = Randomizer.getRandom();
     // A shared male birth rate for animals of this species
-    private static final int maleBirthRate;
+    private static final float maleBirthRate;
     
     // Individual characteristics (instance fields).
